requirements.txt
Django==5.2.1
dj-database-url==3.0.1
django-environ==0.12.0
django-allauth==65.10.0
django-htmx==1.23.2
sqlparse==0.5.3
channels==4.2.2
daphne==4.2.1
whitenoise==6.9.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
pillow==11.0.0
requests==2.32.4
humanize==4.9.0
shortuuid==1.0.13
PyJWT==2.10.1
